通信功能 1.连接与断开服务端 2.应用进入后台断开连接 3.应用重新进入前台连接恢复

+++ skyInnerSX/skyInnerSX/Popover Views/Settings Popover View/skySettingConnectionVC.h
@@ -25,7 +25,7 @@
 @protocol skySettingConnectionVCDelegate <NSObject>
 
 // 连接控制器
-- (void)connectToController:(NSString *)ipAddress andPort:(NSInteger)nPort;
+- (void)connectToController:(NSString *)ipAddress andPort:(int)nPort;
 // 断开控制器
 - (void)disconnectController;
